+++ aim/web/ui/src/pages/Metrics/Metrics.tsx
@@ -234,7 +234,7 @@ function Metrics(
                               ? 'medium'
                               : 'large'
                           }
-                          sortOptions={props.groupingSelectOptions}
+                          sortOptions={props.sortOptions}
                           sortFields={props.sortFields}
                           hiddenRows={props.hiddenMetrics}
                           hiddenColumns={props.hiddenColumns}